+++ app/Http/routes.js
@@ -19,7 +19,8 @@ Route.get('/g-auth/result', 'GAuthController.handleCallback');
 
 // See 'Reourceful Routes' @ http://www.adonisjs.com/docs/3.1/routing
 // These API routes were pre-vue and will need to be modified
-Route.resources('users', 'UserController');
+Route.resources('users', 'UserController')
+  .only('show', 'update');
 Route.resources('submissions', 'SubmissionController');
 
 /*