+++ app/Http/Middleware/AttachTokenHeader.js
@@ -1,13 +1,15 @@
 'use strict';
 
-const Token = use('App/Model/Token');
+const User = use('App/Model/User');
 
 class AttachTokenHeader {
 
   * handle (request, response, next) {
-    const user = request.cookie('user', null);
-    const token = yield Token.findBy('user_id', user);
-    response.header('Authorization', 'Bearer ' + token.token);
+    const userId = request.cookie('user', null);
+    const user = yield User.findBy('email', userId);
+    user.id = userId;
+    const token = yield request.auth.authenticator('api').generate(user)
+    request.header('Authorization', 'Bearer ' + token.token);
     yield next;
   }