+++ resources/src/router/index.js
@@ -1,10 +1,10 @@
-import Vue from 'vue'
-import Router from 'vue-router'
+import Vue from 'vue';
+import Router from 'vue-router';
 
-Vue.use(Router)
+Vue.use(Router);
 
-import Home from '../views/Home.vue'
-import About from '../views/About.vue'
+import Home from '../views/Home.vue';
+import About from '../views/About.vue';
 
 export default new Router({
   mode: 'history',
@@ -13,4 +13,4 @@ export default new Router({
     { path: '/about', component: About },
     { path: '*', component: Home }
   ]
-})
+});