+++ resources/src/server-entry.js
@@ -1,6 +1,6 @@
-import { app, router, store } from './app'
+import { app, router, store } from './app';
 
-const isDev = process.env.NODE_ENV !== 'production'
+const isDev = process.env.NODE_ENV !== 'production';
 
 // This exported function will be called by `bundleRenderer`.
 // This is where we perform data-prefetching to determine the
@@ -9,9 +9,9 @@ const isDev = process.env.NODE_ENV !== 'production'
 // return a Promise that resolves to the app instance.
 export default context => {
   // set router's location
-  router.push(context.url)
+  router.push(context.url);
 
-  const s = isDev && Date.now()
+  const s = isDev && Date.now();
 
   // Call preFetch hooks on components matched by the route.
   // A preFetch hook dispatches a store action and returns a Promise,
@@ -19,17 +19,17 @@ export default context => {
   // updated.
   return Promise.all(router.getMatchedComponents().map(component => {
     if (component.preFetch) {
-      return component.preFetch(store)
+      return component.preFetch(store);
     }
   })).then(() => {
-    isDev && console.log(`data pre-fetch: ${Date.now() - s}ms`)
+    isDev && console.log(`data pre-fetch: ${Date.now() - s}ms`);
     // After all preFetch hooks are resolved, our store is now
     // filled with the state needed to render the app.
     // Expose the state on the render context, and let the request handler
     // inline the state in the HTML response. This allows the client-side
     // store to pick-up the server-side state without having to duplicate
     // the initial data fetching on the client.
-    context.initialState = store.state
-    return app
-  })
-}
+    context.initialState = store.state;
+    return app;
+  });
+};