+++ resources/src/store/index.js
@@ -1,7 +1,7 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
+import Vue from 'vue';
+import Vuex from 'vuex';
 
-Vue.use(Vuex)
+Vue.use(Vuex);
 
 const store = new Vuex.Store({
   state: {},
@@ -11,6 +11,6 @@ const store = new Vuex.Store({
   mutations: {},
 
   getters: {}
-})
+});
 
-export default store
+export default store;