+++ CHANGELOG.md
@@ -7,12 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 ### Added
 - User, Submission, and Profile routes
-- Implement google auth
-- Login/Create users
-- Route logged in user to profile
+- Implement google auth (partial)
+- Create users
+- Integrate route auth (partial)
 
 ### Changed
 - Use ace migration tools to refactor tables
+- Disable route auth for view dev
 
 ## 0.2.2 - 2016-11-8
 ### Added